Speed measure

+++ bicycle.h
@@ -5,9 +5,10 @@
  *      Author: JVillanueva
  */
 
-#include "ftm_freq.h"
 #include "graphic_interface.h"
 #include "MPU6050.h"
+#include "rtc_mod.h"
+#include "freq.h"
 
 #define WHEEL       1.95f
 
@@ -26,7 +27,7 @@ void bicyclye_init_modules(void);
 
 void bicycle_update_FSM(void);
 
-float bicycle_calculate_speed(uint32_t freq);
+float bicycle_calculate_speed(float freq);
 
 uint32_t bicycle_calculate_distance(float speed);
